# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vibe80 REST API",
    "version": "1.0.0",
    "description": "Spec (workspace auth, providers at workspace level)"
  },
  "servers": [
    {
      "url": "http://localhost:5179"
    }
  ],
  "paths": {
...
    "/api/sessions/{sessionId}/worktrees": {
      "get": {
        "summary": "List worktrees",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Worktree list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorktreeListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "post": {
        "summary": "Create worktree",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorktreeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Worktree created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorktreeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ]
      }
    },
    "/api/sessions/{sessionId}/worktrees/{worktreeId}": {
      "get": {
        "summary": "Get worktree",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Worktree detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorktreeDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Delete worktree",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Worktree removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "summary": "Rename worktree",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameWorktreeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Worktree updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorktreeRenameResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/worktrees/{worktreeId}/file": {
      "get": {
        "summary": "Read file",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/FilePath"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "File content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileReadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "summary": "Write file",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Write ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileWriteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/worktrees/{worktreeId}/status": {
      "get": {
        "summary": "Worktree git status",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Status entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorktreeStatusResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/worktrees/{worktreeId}/diff": {
      "get": {
        "summary": "Worktree diff",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Worktree diff",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoDiff"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/worktrees/{worktreeId}/commits": {
      "get": {
        "summary": "Worktree commits",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/CommitsLimit"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Commit list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/worktrees/{worktreeId}/merge": {
      "post": {
        "summary": "Merge worktree into target worktree",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeWorktreeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Merge result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeWorktreeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/worktrees/{worktreeId}/abort-merge": {
      "post": {
        "summary": "Abort merge",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionOnlyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Aborted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/worktrees/{worktreeId}/cherry-pick": {
      "post": {
        "summary": "Cherry-pick commit",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CherryPickRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cherry-pick result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeWorktreeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/sessions/{sessionId}/worktrees/{worktreeId}/browse": {
      "get": {
        "summary": "Browse worktree directory",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/BrowsePath"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Directory entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryBrowseResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/worktrees/{worktreeId}/messages": {
      "get": {
        "summary": "Get worktree messages",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorktreeId"
          },
          {
            "$ref": "#/components/parameters/MessagesLimit"
          },
          {
            "$ref": "#/components/parameters/BeforeMessageId"
          },
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Worktree messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorktreeMessagesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "SessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^s[0-9a-f]{24}$"
        }
      },
      "WorktreeId": {
        "name": "worktreeId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "WorktreeListResponse": {
        "type": "object",
        "properties": {
          "worktrees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorktreeSummary"
            }
          }
        },
        "required": [
          "worktrees"
        ],
        "additionalProperties": false
      },
      "CreateWorktreeRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/Provider"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "parentWorktreeId": {
            "type": [
              "string",
              "null"
            ]
          },
          "startingBranch": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "reasoningEffort": {
            "type": [
              "string",
              "null"
            ]
          },
          "internetAccess": {
            "type": "boolean"
          },
          "denyGitCredentialsAccess": {
            "type": "boolean"
          }
        },
        "required": [
          "provider"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}